Extract guide block view holders into separate file
